🛠️ [FIX] Reply to a comment. 1) Make this workable. 2) Restrict user reply to only parent comment, not reply to a reply.
